+++ Runtime/EColor.cs
@@ -5,17 +5,35 @@ namespace SaintsField
 {
     public enum EColor
     {
-        Default,
-        Clear,
-        White,
+        // https://docs.unity3d.com/Packages/com.unity.ugui@1.0/manual/StyledText.html#ColorNames
+        // this this rich text
+        Aqua,
         Black,
-        Gray,
-        Red,
-        Pink,
-        Orange,
-        Yellow,
-        Green,
         Blue,
+        Brown,
+        Cyan,
+        DarkBlue,
+        Fuchsia,
+        Green,
+        Gray,
+        Grey,
+        LightBlue,
+        Lime,
+        Magenta,
+        Maroon,
+        Navy,
+        Olive,
+        Orange,
+        Purple,
+        Red,
+        Silver,
+        Teal,
+        White,
+        Yellow,
+
+        // this is extended from [NaughtyAttributes](https://github.com/dbrizov/NaughtyAttributes/blob/master/Assets/NaughtyAttributes/Scripts/Core/Utility/EColor.cs)
+        Clear,
+        Pink,
         Indigo,
         Violet,
     }
@@ -26,26 +44,56 @@ namespace SaintsField
         {
             switch (color)
             {
-                case EColor.Clear:
-                    return Color.clear;
-                case EColor.White:
-                    return Color.white;
+                case EColor.Aqua:
+                case EColor.Cyan:
+                    return Color.cyan;
                 case EColor.Black:
                     return Color.black;
-                case EColor.Gray:
-                    return Color.gray;
-                case EColor.Red:
-                    return Color.red;
-                case EColor.Pink:
-                    return Colors.GetColorByName("pink");
-                case EColor.Orange:
-                    return Colors.GetColorByName("orange");
-                case EColor.Yellow:
-                    return Color.yellow;
-                case EColor.Green:
-                    return Color.green;
                 case EColor.Blue:
                     return Color.blue;
+                case EColor.Brown:
+                    return Colors.GetColorByName("brown");
+                case EColor.DarkBlue:
+                    return Colors.GetColorByName("darkblue");
+                case EColor.Fuchsia:
+                    return Color.magenta;
+                case EColor.Green:
+                    return Colors.GetColorByName("green");
+                case EColor.Gray:
+                case EColor.Grey:
+                    return Color.gray;
+                case EColor.LightBlue:
+                    return Colors.GetColorByName("lightblue");
+                case EColor.Lime:
+                    return Color.green;
+                case EColor.Magenta:
+                    return Color.magenta;
+                case EColor.Maroon:
+                    return Colors.GetColorByName("maroon");
+                case EColor.Navy:
+                    return Colors.GetColorByName("navy");
+                case EColor.Olive:
+                    return Colors.GetColorByName("olive");
+                case EColor.Orange:
+                    return Colors.GetColorByName("orange");
+                case EColor.Purple:
+                    return Colors.GetColorByName("purple");
+                case EColor.Red:
+                    return Color.red;
+                case EColor.Silver:
+                    return Colors.GetColorByName("silver");
+                case EColor.Teal:
+                    return Colors.GetColorByName("teal");
+                case EColor.White:
+                    return Color.white;
+                case EColor.Yellow:
+                    return Color.yellow;
+
+                case EColor.Clear:
+                    return Color.clear;
+
+                case EColor.Pink:
+                    return Colors.GetColorByName("pink");
                 case EColor.Indigo:
                     return Colors.GetColorByName("indigo");
                 case EColor.Violet: