+++ Runtime/Utils/Colors.cs
@@ -6,6 +6,11 @@ namespace SaintsField.Utils
     {
         public static Color GetColorByStringPresent(string name)
         {
+            if (name == null)
+            {
+                return Color.white;
+            }
+
             if(ColorUtility.TryParseHtmlString(name, out Color color))
             {
                 return color;
@@ -21,17 +26,19 @@ namespace SaintsField.Utils
             //     ? (Color32)reflectValue
             //     : Color.white;
 
+            // Debug.Log($"colorName={name}");
+
             return GetColorByName(name);
         }
 
         public static Color GetColorByName(string name)
         {
-            switch (name)
+            switch (name.ToLower())
             {
                 case "red":
                     return Color.red;
-                case "green":
-                    return Color.green;
+                case "green":  // this is not error:  https://docs.unity3d.com/Packages/com.unity.ugui@1.0/manual/StyledText.html#ColorNames
+                    return new Color(0, 128, 0, 255);
                 case "blue":
                     return Color.blue;
                 case "white":
@@ -40,12 +47,13 @@ namespace SaintsField.Utils
                     return Color.black;
                 case "yellow":
                     return Color.yellow;
+                case "aqua":
                 case "cyan":
                     return Color.cyan;
+                case "fuchsia":
                 case "magenta":
                     return Color.magenta;
                 case "gray":
-                    return Color.gray;
                 case "grey":
                     return Color.grey;
                 case "clear":
@@ -53,36 +61,57 @@ namespace SaintsField.Utils
                 case "pink":
                     return new Color32(255, 152, 203, 255);
                 case "orange":
-                    return new Color32(255, 128, 0, 255);
+                    return new Color32(255, 165, 0, 255);
                 case "indigo":
                     return new Color32(75, 0, 130, 255);
                 case "violet":
                     return new Color32(128, 0, 255, 255);
+                case "brown":
+                    return new Color32(165, 42, 42, 255);
+                case "darkblue":
+                    return new Color32(0, 0, 160, 255);
+                case "lightblue":
+                    return new Color32(173, 216, 230, 255);
+                case "lime":
+                    return Color.green;
+                case "maroon":
+                    return new Color32(128, 0, 0, 255);
+                case "navy":
+                    return new Color32(0, 0, 128, 255);
+                case "olive":
+                    return new Color32(128, 128, 0, 255);
+                case "purple":
+                    return new Color32(128, 0, 128, 255);
+                case "silver":
+                    return new Color32(192, 192, 192, 0);
+                case "teal":
+                    return new Color32(0, 128, 128, 255);
+
                 default:
                     return Color.white;
             }
         }
 
-        public static bool ColorNameSupported(string name)
-        {
-            switch (name)
-            {
-                case "red":
-                case "green":
-                case "blue":
-                case "white":
-                case "black":
-                case "yellow":
-                case "cyan":
-                case "magenta":
-                case "gray":
-                case "grey":
-                case "clear":
-                    return true;
-                default:
-                    return false;
-            }
-        }
+        // public static bool ColorNameSupported(string name)
+        // {
+        //     switch (name.ToLower())
+        //     {
+        //         case "red":
+        //         case "green":
+        //         case "blue":
+        //         case "white":
+        //         case "black":
+        //         case "yellow":
+        //         case "cyan":
+        //         case "magenta":
+        //         case "gray":
+        //         case "grey":
+        //         case "clear":
+        //             return true;
+        //         default:
+        //             return false;
+        //     }
+        // }
 
         public static string ToHtmlHexString(Color color)
         {